requirements.txt
google-generativeai>=0.8.0
langchain-google-genai>=1.0.3
langchain-core>=0.2.36
pydantic>=2.5.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
streamlit>=1.28.0
rich>=13.7.0
